fix: make props form use field order
