+++ libs/frontend/domain/type/src/shared/index.ts
@@ -1,2 +1,3 @@
+export * from './sortFields'
 export * from './TypeSelect'
 export * from './typeSelectOptions'
